requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.2.3
